+++ cli/src/transactions.rs
@@ -86,7 +86,7 @@ pub fn operations(client: &Client, matches: &ArgMatches) -> Result<()> {
         .value_of("Hash")
         .expect("Transaction identifier hash is required");
 
-    let endpoint = transaction::Payments::new(hash);
+    let endpoint = transaction::Operations::new(hash);
     let endpoint = pager.assign(endpoint);
     let endpoint = cursor::assign_from_arg(matches, endpoint);
     let endpoint = ordering::assign_from_arg(matches, endpoint);
@@ -98,6 +98,31 @@ pub fn operations(client: &Client, matches: &ArgMatches) -> Result<()> {
         Ok(op) => {
             println!("ID:   {}", op.id());
             println!("Type: {}", op.kind_name());
+            println!();
+        }
+        Err(err) => res = Err(err),
+    });
+    res
+}
+
+pub fn effects(client: &Client, matches: &ArgMatches) -> Result<()> {
+    let pager = Pager::from_arg(&matches);
+    let hash = matches
+        .value_of("Hash")
+        .expect("Transaction identifier hash is required");
+
+    let endpoint = transaction::Effects::new(hash);
+    let endpoint = pager.assign(endpoint);
+    let endpoint = cursor::assign_from_arg(matches, endpoint);
+    let endpoint = ordering::assign_from_arg(matches, endpoint);
+
+    let iter = sync::Iter::new(&client, endpoint);
+
+    let mut res = Ok(());
+    pager.paginate(iter, |result| match result {
+        Ok(effect) => {
+            println!("ID:   {}", effect.id());
+            println!("Type: {}", effect.kind_name());
             println!();
         }
         Err(err) => res = Err(err),