+++ resources/src/effect/account/home_domain_updated.rs
@@ -1,6 +1,6 @@
 /// This effect can be the result of a set options operation and represents
 /// the fact that an account's home domain has changed
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, Clone)]
 pub struct HomeDomainUpdated {
     account: String,
     home_domain: String,