+++ resources/src/effect/signer/created.rs
@@ -1,6 +1,6 @@
 /// This effect can be the result of a set options operation and represents
 /// the fact that a new signer has been created for an account.
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, Clone)]
 pub struct Created {
     account: String,
     public_key: String,