+++ resources/src/effect/trustline/authorized.rs
@@ -1,7 +1,7 @@
 use asset::AssetIdentifier;
 /// This effect can be the result of a allow trust operation and represents
 /// the fact that an asset issuer will allow an account to hold its assets.
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, Clone)]
 pub struct Authorized {
     account: String,
     asset: AssetIdentifier,