+++ cli/CHANGELOG.md
@@ -12,3 +12,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Transaction payments can be seen
 - Transaction details can be seen
 - All ledgers can be seen
+- Transaction operations can be seen
+- Transaction effects can be seen